Move new code to netlink-conky
Replace define with a wrapper struct.
